+++ gridopt/power_flow/ac_opf.py
@@ -27,7 +27,7 @@ class ACOPF(PFmethod):
                    'weight_b' : 0.,         # weight for shunt susceptances regularization
                    'thermal_limits': False, # flag for thermal limits
                    'vmin_thresh': 0.1,      # threshold for vmin termination
-                   'optsolver': 'augl'}     # OPTALG optimization solver (augl,ipopt)
+                   'solver': 'augl'}        # OPTALG optimization solver (augl, ipopt, inlp)
 
     _parameters_augl = {'feastol' : 1e-4,
                         'optol' : 1e-4,
@@ -44,7 +44,7 @@ class ACOPF(PFmethod):
         # Parent init
         PFmethod.__init__(self)
 
-        # Optsolver params
+        # Solver params
         augl_params = OptSolverAugL.parameters.copy()
         augl_params.update(self._parameters_augl)   # overwrite defaults
 
@@ -55,9 +55,9 @@ class ACOPF(PFmethod):
         inlp_params.update(self._parameters_inlp)   # overwrite defaults
 
         self._parameters.update(ACOPF._parameters)
-        self._parameters['optsolver_parameters'] = {'augl': augl_params,
-                                                    'ipopt': ipopt_params,
-                                                    'inlp': inlp_params}
+        self._parameters['solver_parameters'] = {'augl': augl_params,
+                                                 'ipopt': ipopt_params,
+                                                 'inlp': inlp_params}
                    
     def create_problem(self,net):
         
@@ -142,19 +142,19 @@ class ACOPF(PFmethod):
         # Parameters
         params = self._parameters
         vmin_thresh = params['vmin_thresh']
-        optsolver_name = params['optsolver']
-        optsolver_params = params['optsolver_parameters']
+        solver_name = params['solver']
+        solver_params = params['solver_parameters']
 
         # Opt solver
-        if optsolver_name == 'augl':
-            optsolver = OptSolverAugL()
-        elif optsolver_name == 'ipopt':
-            optsolver = OptSolverIpopt()
-        elif optsolver_name == 'inlp':
-            optsolver = OptSolverINLP()
+        if solver_name == 'augl':
+            solver = OptSolverAugL()
+        elif solver_name == 'ipopt':
+            solver = OptSolverIpopt()
+        elif solver_name == 'inlp':
+            solver = OptSolverINLP()
         else:
             raise PFmethodError_BadOptSolver()
-        optsolver.set_parameters(optsolver_params[optsolver_name])
+        solver.set_parameters(solver_params[solver_name])
 
         # Copy network
         net = net.get_copy()
@@ -170,17 +170,17 @@ class ACOPF(PFmethod):
                 return True
             else:
                 return False
-        optsolver.add_termination(OptTermination(t1,'low voltage'))
+        solver.add_termination(OptTermination(t1,'low voltage'))
         
         # Info printer
         info_printer = self.get_info_printer()
-        optsolver.set_info_printer(info_printer)
+        solver.set_info_printer(info_printer)
         
         # Solve
         update = True
         t0 = time.time()
         try:
-            optsolver.solve(problem)
+            solver.solve(problem)
         except OptSolverError as e:
             raise PFmethodError_SolverError(e)
         except Exception as e:
@@ -190,19 +190,19 @@ class ACOPF(PFmethod):
             
             # Update network
             if update:
-                net.set_var_values(optsolver.get_primal_variables()[:net.num_vars])
+                net.set_var_values(solver.get_primal_variables()[:net.num_vars])
                 net.update_properties()
                 net.clear_sensitivities()
-                problem.store_sensitivities(*optsolver.get_dual_variables())
+                problem.store_sensitivities(*solver.get_dual_variables())
 
             # Save results
-            self.set_solver_name(optsolver_name)
-            self.set_solver_status(optsolver.get_status())
-            self.set_solver_message(optsolver.get_error_msg())
-            self.set_solver_iterations(optsolver.get_iterations())
+            self.set_solver_name(solver_name)
+            self.set_solver_status(solver.get_status())
+            self.set_solver_message(solver.get_error_msg())
+            self.set_solver_iterations(solver.get_iterations())
             self.set_solver_time(time.time()-t0)
-            self.set_solver_primal_variables(optsolver.get_primal_variables())
-            self.set_solver_dual_variables(optsolver.get_dual_variables())
+            self.set_solver_primal_variables(solver.get_primal_variables())
+            self.set_solver_dual_variables(solver.get_dual_variables())
             self.set_problem(None) # skip for now
             self.set_problem_time(problem_time)
             self.set_network_snapshot(net)